welding attempt JOHN
a1mess3-4 can be combined with RMRCircularpath to create the welding sequence, it can be repeated as much as needed like a production line.

i cant htink anymore but i know it can be done.

a pause can be used to let the ur3e do its thing while the kuka simply holds the flange in place